Added Watir web driver support
