+++ lib/billy/capybara/cucumber.rb
@@ -1,0 +1,3 @@
+require 'capybara/cucumber'
+require 'billy/capybara/capybara'
+require 'billy/init/cucumber'
