+++ lib/billy/capybara/railtie.rb
@@ -1,0 +1,2 @@
+require 'billy/capybara/capybara'
+require 'billy/init/railtie'
