+++ lib/billy/capybara/rspec.rb
@@ -1,0 +1,3 @@
+require 'capybara/rspec'
+require 'billy/capybara/capybara'
+require 'billy/init/rspec'
