+++ lib/billy/cucumber.rb
@@ -1,19 +1,6 @@
-require 'cucumber'
+# Deprecated
 require 'capybara/cucumber'
-require 'billy'
+require 'billy/capybara/capybara'
+require 'billy/init/cucumber'
 
-Billy.register_drivers
-
-module Billy
-  module CucumberHelper
-    def proxy
-      Billy.proxy
-    end
-  end
-end
-
-World(Billy::CucumberHelper)
-
-After('@billy') do
-  proxy.reset
-end
+warn "[DEPRECATION] `require 'billy/cucumber'` is deprecated. Please use `require 'billy/capybara/cucumber'` instead."
