+++ lib/billy/init/cucumber.rb
@@ -1,0 +1,17 @@
+require 'cucumber'
+
+Billy.register_drivers
+
+module Billy
+  module CucumberHelper
+    def proxy
+      Billy.proxy
+    end
+  end
+end
+
+World(Billy::CucumberHelper)
+
+After('@billy') do
+  proxy.reset
+end
