+++ lib/billy/init/railtie.rb
@@ -1,0 +1,9 @@
+module Billy
+  class Railtie < Rails::Railtie
+    railtie_name 'billy'
+
+    rake_tasks do
+      load 'tasks/billy.rake'
+    end
+  end
+end
