+++ lib/billy/watir/rspec.rb
@@ -1,0 +1,2 @@
+require 'billy/watir/watir'
+require 'billy/init/rspec'
