+++ spec/spec_helper.rb
@@ -1,7 +1,7 @@
 Dir[File.expand_path('../support/**/*.rb', __FILE__)].each { |f| require f }
 
 require 'pry'
-require 'billy/rspec'
+require 'billy/capybara/rspec'
 require 'rack'
 require 'logger'
